# test: speed up some tests + minor refactorings