--- a/releasenotes/notes/add-huggingface-api-text-embeddings-inference-to-component-rankers-0e3f54e523e42141.yaml
+++ b/releasenotes/notes/add-huggingface-api-text-embeddings-inference-to-component-rankers-0e3f54e523e42141.yaml
@@ -2,5 +2,4 @@
 features:
   - |
     Added new `HuggingFaceTEIRanker` component to enable reranking with Text Embeddings Inference (TEI) API.
-    This component supports both both self-hosted Text Embeddings Inference services and Hugging Face Inference
-    Endpoints.
+    This component supports both self-hosted Text Embeddings Inference services and Hugging Face Inference Endpoints.