--- a/test/components/agents/test_agent.py
+++ b/test/components/agents/test_agent.py
@@ -804,8 +804,9 @@ class TestAgent:
         assert isinstance(result["last_message"], ChatMessage)
         assert result["messages"][-1] == result["last_message"]
 
+    @pytest.mark.integration
     @pytest.mark.skipif(not os.environ.get("OPENAI_API_KEY"), reason="OPENAI_API_KEY not set")
-    def test_agent_streaming_with_tool_call(self, monkeypatch, weather_tool):
+    def test_agent_streaming_with_tool_call(self, weather_tool):
         chat_generator = OpenAIChatGenerator()
         agent = Agent(chat_generator=chat_generator, tools=[weather_tool])
         agent.warm_up()