… test/components/embedders/types/test_protocol.py