### Related Issues

On my Mac, I ran a command to identify slow tests on main: `hatch run test:unit -vvv --durations=10 --durations-min=0.5 test`

```
2.02s call     test/components/fetchers/test_link_content_fetcher.py::TestLinkContentFetcherAsync::test_run_async_user_agent_rotation
1.90s call     test/components/agents/test_agent.py::TestAgent::test_agent_streaming_with_tool_call
1.05s call     test/utils/test_requests_utils.py::TestAsyncRequestWithRetry::test_async_request_with_retry_retries_on_status_code
1.04s call     test/utils/test_requests_utils.py::TestAsyncRequestWithRetry::test_async_request_with_retry_retries_on_error
1.01s call     test/utils/test_requests_utils.py::TestRequestWithRetry::test_request_with_retry_retries_on_status_code
1.01s call     test/utils/test_requests_utils.py::TestRequestWithRetry::test_request_with_retry_retries_on_error
0.50s call     test/components/agents/test_agent.py::TestAgent::test_output_types
0.39s call     test/components/generators/chat/test_openai_async.py::TestOpenAIChatGeneratorAsync::test_run_with_wrong_model_async
0.31s call     test/core/pipeline/test_pipeline_base.py::TestPipelineBase::test_show_in_notebook
0.30s call     test/components/preprocessors/test_sentence_tokenizer.py::test_split_sentences_performance
```

After this PR

```
0.49s call     test/components/agents/test_agent.py::TestAgent::test_output_types
0.41s call     test/components/generators/chat/test_openai_async.py::TestOpenAIChatGeneratorAsync::test_run_with_wrong_model_async
0.31s call     test/core/pipeline/test_pipeline_base.py::TestPipelineBase::test_show_in_notebook
0.29s call     test/components/preprocessors/test_sentence_tokenizer.py::test_split_sentences_performance
0.26s call     test/components/builders/test_chat_prompt_builder.py::TestChatPromptBuilder::test_invalid_timezone
0.21s call     test/components/classifiers/test_document_language_classifier.py::TestDocumentLanguageClassifier::test_detect_language
0.15s call     test/document_stores/test_in_memory.py::TestMemoryDocumentStore::test_executor_shutdown
0.15s call     test/components/generators/chat/test_hugging_face_local.py::TestHuggingFaceLocalChatGenerator::test_executor_shutdown
0.13s setup    test/components/agents/test_agent.py::TestAgent::test_run_with_params_streaming
0.07s call     test/components/converters/test_pdfminer_to_document.py::TestPDFMinerToDocument::test_run_doc_metadata
```

### Proposed Changes:
- speed up slow tests involving `tenacity` by mocking `sleep`/`asyncio.sleep` to not actually wait
- marked `test_agent_streaming_with_tool_call` as integration test
- remove `test_protocol.py`: (not slow) we attempted to test the `TextEmbedder` Protocol, which is a static typing abstraction and not intended for runtime testing with pytest (this specific Protocol is not runtime checkable)
- fix typo in release note

### How did you test it?
CI

### Checklist

- I have read the [contributors guidelines](https://github.com/deepset-ai/haystack/blob/main/CONTRIBUTING.md) and the [code of conduct](https://github.com/deepset-ai/haystack/blob/main/code_of_conduct.txt)
- I have updated the related issue with new insights and changes
- I added unit tests and updated the docstrings
- I've used one of the [conventional commit types](https://www.conventionalcommits.org/en/v1.0.0/) for my PR title: `fix:`, `feat:`, `build:`, `chore:`, `ci:`, `docs:`, `style:`, `refactor:`, `perf:`, `test:` and added `!` in case the PR includes breaking changes.
- I documented my code
- I ran [pre-commit hooks](https://github.com/deepset-ai/haystack/blob/main/CONTRIBUTING.md#installation) and fixed any issue