@anakin87 2025-05-29
> @anakin87 looks good but why does the release [note](https://github.com/deepset-ai/haystack/pull/9451/files#diff-b2036be962f0708b1746b1c76d3dc81893054964b01bf06e38e629946cb6c931) talk about TEI rankers

I'm just fixing a typo in that release note. "both both"